+++ src/App.tsx
@@ -15,7 +15,7 @@ import HomeMyworkPage from '@/page/Home/HomeMyworkPage';
 import HomePage from '@/page/Home/HomePage';
 import SocialHomePage from '@/page/HomeSocial/SocialHomePage';
 import { CareGiverInfoPage } from '@/page/Matching/CareGiverInfoPage';
-import { MatchingApplyPage } from '@/page/Matching/MatchingApplyPage';
+
 import { MatchingElderPage } from '@/page/Matching/MatchingElderPage';
 import { MatchingInformationPage } from '@/page/Matching/MatchingInformationPage';
 import MatchingStatus from '@/page/Matching/MatchingStatus';
@@ -29,6 +29,7 @@ import { OnboardingPage } from '@/page/Onboarding/OnboardingPage';
 import { CommunitySignUpPage } from '@/page/SignUp/CommunitySignUpPage';
 import { InstitutionSignUpPage } from '@/page/SignUp/InstitutionSignUpPage';
 import { SignUpPage } from '@/page/SignUp/SignUpPage';
+import { SocialWorkerMatchingPage } from '@/page/SocialWorkerMatching/SocialWorkerMatchingPage';
 import SplashPage from '@/page/SplashPage';
 import { TestPage } from '@/page/TestPage';
 import { WorkDetailPage } from '@/page/Works/WorkDetailPage';
@@ -85,7 +86,7 @@ function App() {
       <Route path="/home/social" element={<SocialHomePage />} />
       <Route path="/elderly" element={<ElderlyPage />} />
       <Route path="/elderly/new" element={<ElderlyRegisterPage />} />
-      <Route path="/matching" element={<MatchingApplyPage />} />
+      <Route path="/match/social" element={<SocialWorkerMatchingPage />} />
       <Route path="/matching/elder-apply" element={<MatchingElderPage />} />
       <Route
         path="/matching/info/:recruitmentId"
